More EL7 fixes
Because EL7 (specifically CentOS 7) doesn't have a ktuned service, I had to reverse the if/else check.  Consequently, reordered the Exec and Service resources to eliminate an otherwise awkward "Service[$tuned_services]" (not guaranteed to work, IIRC) or having to duplicate the if check for when the ktune service does/doesn't exist.  HTH! ;)